Basic timetravel debugger (reverse step) implementation.
Note: RAM of XSM is not restored during reverse step.
Full record and replay of RAM takes up too much memory and freezes
system.
Differential record and replay works, but is too slow to be practical.
Basic version of the differential version is in place, commented out.
You are welcome to tinker with it.

Copying over the memory and iterating over it to diff it with previous
copy takes O(n) time.
Try changing the memory data structor from an array to some form of balanced tree,
where node->data is a stack implemented as a linked list of the struct {step_no, value} where
step_no is the number of step since start of the simulator and value is
the changed value.
Then during reverse step, if current step_no == node->step_no, then pop
off the last element from stack.
This is O(1) during forward step, but O(n) during reverse step, which is
reasonable.